backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.5.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
phonenumbers==8.13.19
google-generativeai==0.3.2
pytest==7.4.2
pytest-flask==1.2.0
pytest-mock==3.11.1
coverage==7.3.1
flasgger==0.9.7.1
gunicorn==21.2.0